+++ Identity_Field/src/main/java/com/iluwater/fieild/model/Book.java
@@ -22,31 +22,22 @@
   * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   * THE SOFTWARE.
 */
-package com.iluwater.fieild.Model;
-/*
- * The DomainObject class provides a base for domain entities
- * that require unique identification within the application.
- *
- * <p>All child classes inherit the unique identifier field
- * and associated functionality.</p>
- */
+
+package com.iluwater.fieild.model;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
 import lombok.Getter;
 import lombok.Setter;
-
-@Getter
 @Setter
-@MappedSuperclass
+@Getter
 @Entity
-public abstract class DomainObject {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY) // automatically generate unique values for primary key columns
-  private Long id;
-  //strategy = GenerationType.IDENTITY generate the primary key value by the database itself using the auto-increment column option
+public class Book extends DomainObject {
+  private String title;
+  private String author;
 
+/** Book constructor*/
+  public Book(String title, String author) {
+    this.title = title;
+    this.author = author;
+  }
 
 }
