+++ Identity_Field/src/main/java/com/iluwater/fieild/services/BookRepository.java
@@ -1,15 +1,17 @@
-package com.iluwater.fieild.Services;
-import com.iluwater.fieild.Model.Book;
-import org.springframework.beans.factory.annotation.Autowired;
+package com.iluwater.fieild.services;
+import com.iluwater.fieild.model.Book;
 import org.springframework.stereotype.Repository;
 import javax.persistence.EntityManager;
 
 @Repository
 public class BookRepository {
-  @Autowired
-  private EntityManager entityManager;
+  private final EntityManager entityManager;
 
-  public Book createBook(String title,String author) {
+  public BookRepository(EntityManager entityManager) {
+    this.entityManager = entityManager;
+  }
+
+  public Book createBook(String title, String author) {
     Book book = new Book(title, author);
     entityManager.persist(book);
     return book;
