+++ Identity_Field/src/test/java/com/iluwater/fieild/ClientTest.java
@@ -24,7 +24,7 @@
  */
 package com.iluwater.fieild;
 
-import com.iluwater.fieild.Model.Client;
+import com.iluwater.fieild.model.Client;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
